@@ -36,14 +36,12 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 
-import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.UUID;
+
 
 public class SignUpActivity extends AppCompatActivity {
     private static final int CAMERA_REQUEST_CODE = 101;
@@ -79,32 +77,28 @@ public class SignUpActivity extends AppCompatActivity {
         appLogo = findViewById(R.id.app_logo);
 
         // Set up button listener
-        signUpButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String name = nameInput.getText().toString().trim();
-                String email = emailInput.getText().toString().trim();
-                String password = passwordInput.getText().toString().trim();
-                String confirmPassword = confirmPasswordInput.getText().toString().trim();
+        signUpButton.setOnClickListener(view -> {
+            String name = nameInput.getText().toString().trim();
+            String email = emailInput.getText().toString().trim();
+            String password = passwordInput.getText().toString().trim();
+            String confirmPassword = confirmPasswordInput.getText().toString().trim();
 
-                // Check if passwords match
-                if (!password.equals(confirmPassword)) {
-                    Toast.makeText(SignUpActivity.this, "Passwords do not match", Toast.LENGTH_SHORT).show();
-                    return;
-                }
+            // Check if passwords match and validate inputs
+            if (!password.equals(confirmPassword)) {
+                Toast.makeText(SignUpActivity.this, "Passwords do not match", Toast.LENGTH_SHORT).show();
+                return;
+            }
 
-                if (validateInput(name, email, password)) {
-                    String hashedPassword = hashPassword(password);
-                    if (hashedPassword != null) {
-                        if (imagePath != null) {
-                            uploadImageToSupabase(name, email, hashedPassword);
-                        } else {
-                            saveToSupabase(name, email, hashedPassword, null);
-                        }
-                    } else {
-                        Log.e(TAG, "Password hashing failed");
-                        Toast.makeText(SignUpActivity.this, "Error hashing password", Toast.LENGTH_SHORT).show();
-                    }
+            // Validate input and proceed if valid
+            if (validateInput(name, email, password)) {
+                String rawPassword = password;                      // keep raw for Supabase Auth
+                String hashedPassword = hashPassword(rawPassword);
+                if (hashedPassword != null) {
+                    // Proceed to check for duplicate entries in the signup table
+                    checkForDuplicateAndSaveUser(name, email, hashedPassword, rawPassword);
+                } else {
+                    Log.e(TAG, "Password hashing failed");
+                    Toast.makeText(SignUpActivity.this, "Error hashing password", Toast.LENGTH_SHORT).show();
                 }
             }
         });
@@ -164,6 +158,168 @@ public class SignUpActivity extends AppCompatActivity {
         });
 
 
+    }
+
+
+
+
+    private void checkForDuplicateAndSaveUser(String name, String email, String password, String rawPassword) {
+        new Thread(() -> {
+            try {
+                OkHttpClient client = new OkHttpClient();
+
+                // Check for duplicate email or username in the signup table
+                JSONObject json = new JSONObject();
+                json.put("emailid", email);
+                json.put("username", name);
+
+                // Create request body to check for duplicates
+                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+
+                // Check in the signup table for duplicate email or username
+                Request request = new Request.Builder()
+                        .url(SUPABASE_URL + "/rest/v1/" + SIGNUP_TABLE + "?emailid=eq." + email + "&username=eq." + name)
+                        .addHeader("apikey", SUPABASE_KEY)
+                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+                        .get()  // Using GET to check if a user with the same email or username exists
+                        .build();
+
+                Response response = client.newCall(request).execute();
+
+                if (response.isSuccessful()) {
+                    String responseBody = response.body() != null ? response.body().string() : "";
+
+                    // If no records found, proceed to save user
+//                    if (responseBody.equals("[]")) {
+//                        String hashedPassword = hashPassword(password);
+//                        if (hashedPassword != null) {
+//                            // Proceed to save to both signup and auth tables
+//                            saveToSupabase(name, email, hashedPassword, null);  // Save to signup table
+//                            sendAuthenticationEmail(email, password);  // Send authentication email
+//                        } else {
+//                            Log.e(TAG, "Password hashing failed");
+//                            runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Error hashing password", Toast.LENGTH_SHORT).show());
+//                        }
+//                    }
+                    if (responseBody.equals("[]")) {
+                        // At this point, the 'password' parameter is already the hashed password
+                        saveToSupabase(name, email, password /* already hashed */, null);
+
+                        // IMPORTANT: Supabase Auth expects the RAW password, not the hash.
+                        // If you want to keep sending the auth email, you must pass the raw password,
+                        // which means you need to have it available here. See note below.
+                        // sendAuthenticationEmail(email, rawPassword);
+                        sendAuthenticationEmail(email, rawPassword);
+                    }
+
+                    else {
+                        // User already exists, show error
+                        runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Username or Email already exists", Toast.LENGTH_SHORT).show());
+                    }
+                } else {
+                    String errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
+                    Log.e(TAG, "Error checking duplicates: " + errorMessage);
+                    runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Error checking duplicates: " + errorMessage, Toast.LENGTH_SHORT).show());
+                }
+            } catch (Exception e) {
+                Log.e(TAG, "Error checking duplicates: " + e.getMessage());
+                e.printStackTrace();
+            }
+        }).start();
+    }
+
+    // Save user data to Supabase signup table
+    private void saveToSupabase(String name, String email, String password, String imageUrl) {
+        new Thread(() -> {
+            try {
+                OkHttpClient client = new OkHttpClient();
+
+                // JSON payload for the signup request
+                JSONObject json = new JSONObject();
+                json.put("username", name);
+                json.put("emailid", email);
+                json.put("password", password);
+
+                // Add image URL to the payload if available
+                if (imageUrl != null) {
+                    json.put("image", imageUrl);
+                }
+
+                // Create request body
+                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+
+                // Build the request to send to Supabase
+                Request request = new Request.Builder()
+                        .url(SUPABASE_URL + "/rest/v1/" + SIGNUP_TABLE)
+                        .addHeader("apikey", SUPABASE_KEY)
+                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+                        .post(requestBody)
+                        .build();
+
+                Response response = client.newCall(request).execute();
+
+                if (response.isSuccessful()) {
+                    Log.d(TAG, "Signup Success");
+                    runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Sign up successful!", Toast.LENGTH_SHORT).show());
+                } else {
+                    String errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
+                    Log.e(TAG, "Signup Failed: " + errorMessage);
+                    runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Signup failed: " + errorMessage, Toast.LENGTH_SHORT).show());
+                }
+            } catch (Exception e) {
+                Log.e(TAG, "Signup request failed: " + e.getMessage());
+                runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Signup failed: " + e.getMessage(), Toast.LENGTH_SHORT).show());
+                e.printStackTrace();
+            }
+        }).start();
+    }
+
+    // Send authentication email after successful signup
+    private void sendAuthenticationEmail(String email, String password) {
+        new Thread(() -> {  // Run the network request on a background thread
+            try {
+                OkHttpClient client = new OkHttpClient();
+
+                // Create the JSON payload for the signup request
+                JSONObject json = new JSONObject();
+                json.put("email", email);
+                json.put("password", password);
+
+                // Set up the request body
+                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+
+                // Call the Supabase Auth API to send the authentication email
+                Request request = new Request.Builder()
+                        .url(SUPABASE_URL + "/auth/v1/signup")
+                        .addHeader("apikey", SUPABASE_KEY)
+                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+                        .post(requestBody)
+                        .build();
+
+                Response response = client.newCall(request).execute();
+
+                // Switch back to the main thread to update UI
+                runOnUiThread(() -> {
+                    if (response.isSuccessful()) {
+                        Log.d(TAG, "Authentication email sent successfully.");
+                        Toast.makeText(SignUpActivity.this, "Authentication email sent", Toast.LENGTH_SHORT).show();
+                    } else {
+                        String errorMessage = null;
+                        try {
+                            errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
+                        } catch (IOException e) {
+                            throw new RuntimeException(e);
+                        }
+                        Log.e(TAG, "Error sending authentication email: " + errorMessage);
+                        Toast.makeText(SignUpActivity.this, "Error sending authentication email: " + errorMessage, Toast.LENGTH_SHORT).show();
+                    }
+                });
+
+            } catch (Exception e) {
+                Log.e(TAG, "Exception occurred while sending authentication email: " + e.getMessage(), e);
+                runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Error sending authentication email: " + e.getMessage(), Toast.LENGTH_SHORT).show());
+            }
+        }).start();  // Start the background thread
     }
 
 
@@ -227,6 +383,105 @@ public class SignUpActivity extends AppCompatActivity {
     }
 
 
+
+
+//    private void sendAuthenticationEmail(String email, String password) {
+//        new Thread(() -> {  // Run the network request on a background thread
+//            try {
+//                OkHttpClient client = new OkHttpClient();
+//
+//                // Create the JSON payload for the signup request
+//                JSONObject json = new JSONObject();
+//                json.put("email", email);
+//                json.put("password", password);
+//
+//                // Set up the request body
+//                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+//
+//                // Call the Supabase Auth API to send the authentication email
+//                Request request = new Request.Builder()
+//                        .url(SUPABASE_URL + "/auth/v1/signup")
+//                        .addHeader("apikey", SUPABASE_KEY)
+//                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+//                        .post(requestBody)
+//                        .build();
+//
+//                Response response = client.newCall(request).execute();
+//
+//                // Switch back to the main thread to update UI
+//                runOnUiThread(() -> {
+//                    if (response.isSuccessful()) {
+//                        Log.d(TAG, "Authentication email sent successfully.");
+//                        Toast.makeText(SignUpActivity.this, "Authentication email sent", Toast.LENGTH_SHORT).show();
+//                    } else {
+//                        String errorMessage = null;
+//                        try {
+//                            errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
+//                        } catch (IOException e) {
+//                            throw new RuntimeException(e);
+//                        }
+//                        Log.e(TAG, "Error sending authentication email: " + errorMessage);
+//                        Toast.makeText(SignUpActivity.this, "Error sending authentication email: " + errorMessage, Toast.LENGTH_SHORT).show();
+//                    }
+//                });
+//
+//            } catch (Exception e) {
+//                Log.e(TAG, "Exception occurred while sending authentication email: " + e.getMessage(), e);
+//                runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Error sending authentication email: " + e.getMessage(), Toast.LENGTH_SHORT).show());
+//            }
+//        }).start();  // Start the background thread
+//    }
+
+
+    private void updateLoginAttempt(String userId) {
+        try {
+            OkHttpClient client = new OkHttpClient();
+
+            // JSON payload to update login_attempt
+            JSONObject json = new JSONObject();
+            json.put("login_attempt", 1);
+
+            // Create the request body
+            RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+
+            // Send the request to update the login_attempt field in the signup table
+            Request request = new Request.Builder()
+                    .url(SUPABASE_URL + "/rest/v1/signup?id=eq." + userId)
+                    .addHeader("apikey", SUPABASE_KEY)
+                    .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+                    .patch(requestBody)
+                    .build();
+
+            Response response = client.newCall(request).execute();
+
+            if (response.isSuccessful()) {
+                Log.d(TAG, "login_attempt updated to 1 for user");
+            } else {
+                Log.e(TAG, "Failed to update login_attempt: " + response.body().string());
+            }
+        } catch (Exception e) {
+            Log.e(TAG, "Error updating login_attempt: " + e.getMessage());
+            e.printStackTrace();
+        }
+    }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
     private void openGallery() {
         Intent galleryIntent = new Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
         startActivityForResult(galleryIntent, GALLERY_REQUEST_CODE);
@@ -272,66 +527,100 @@ public class SignUpActivity extends AppCompatActivity {
         }
     }
 
-    private void saveToSupabase(String name, String email, String password, String imageUrl) {
-        new Thread(() -> {
-            try {
-                OkHttpClient client = new OkHttpClient();
+//    private void saveToSupabase(String name, String email, String password, String imageUrl) {
+//        new Thread(() -> {
+//            try {
+//                OkHttpClient client = new OkHttpClient();
+//
+//                // JSON payload for the signup request
+//                JSONObject json = new JSONObject();
+//                json.put("username", name);
+//                json.put("emailid", email);
+//                json.put("password", password);
+//
+//                // Add image URL to the payload if available
+//                if (imageUrl != null) {
+//                    json.put("image", imageUrl);
+//                }
+//
+//                // Create request body
+//                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+//
+//                // Build the request to send to Supabase
+//                Request request = new Request.Builder()
+//                        .url(SUPABASE_URL + "/rest/v1/" + SIGNUP_TABLE)
+//                        .addHeader("apikey", SUPABASE_KEY)
+//                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+//                        .post(requestBody)
+//                        .build();
+//
+//                Response response = client.newCall(request).execute();
+//
+//                if (response.isSuccessful()) {
+//                    Log.d(TAG, "Signup Success");
+//                    runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Sign up successful!", Toast.LENGTH_SHORT).show());
+//                } else {
+//                    String errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
+//                    Log.e(TAG, "Signup Failed: " + errorMessage);
+//                    runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Signup failed: " + errorMessage, Toast.LENGTH_SHORT).show());
+//                }
+//            } catch (Exception e) {
+//                Log.e(TAG, "Signup request failed: " + e.getMessage());
+//                runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Signup failed: " + e.getMessage(), Toast.LENGTH_SHORT).show());
+//                e.printStackTrace();
+//            }
+//        }).start();
+//    }
+//
+//    // Method to send the authentication email to the user
+//    private void sendAuthenticationEmail(String email, String password) {
+//        new Thread(() -> {  // Run the network request on a background thread
+//            try {
+//                OkHttpClient client = new OkHttpClient();
+//
+//                // Create the JSON payload for the signup request
+//                JSONObject json = new JSONObject();
+//                json.put("email", email);
+//                json.put("password", password);
+//
+//                // Set up the request body
+//                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
+//
+//                // Call the Supabase Auth API to send the authentication email
+//                Request request = new Request.Builder()
+//                        .url(SUPABASE_URL + "/auth/v1/signup")
+//                        .addHeader("apikey", SUPABASE_KEY)
+//                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
+//                        .post(requestBody)
+//                        .build();
+//
+//                Response response = client.newCall(request).execute();
+//
+//                // Switch back to the main thread to update UI
+//                runOnUiThread(() -> {
+//                    if (response.isSuccessful()) {
+//                        Log.d(TAG, "Authentication email sent successfully.");
+//                        Toast.makeText(SignUpActivity.this, "Authentication email sent", Toast.LENGTH_SHORT).show();
+//                    } else {
+//                        String errorMessage = null;
+//                        try {
+//                            errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
+//                        } catch (IOException e) {
+//                            throw new RuntimeException(e);
+//                        }
+//                        Log.e(TAG, "Error sending authentication email: " + errorMessage);
+//                        Toast.makeText(SignUpActivity.this, "Error sending authentication email: " + errorMessage, Toast.LENGTH_SHORT).show();
+//                    }
+//                });
+//
+//            } catch (Exception e) {
+//                Log.e(TAG, "Exception occurred while sending authentication email: " + e.getMessage(), e);
+//                runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Error sending authentication email: " + e.getMessage(), Toast.LENGTH_SHORT).show());
+//            }
+//        }).start();  // Start the background thread
+//    }
 
-                // JSON payload
-                JSONObject json = new JSONObject();
-                json.put("username", name);
-                json.put("emailid", email);
-                json.put("password", password);
 
-                // Add image URL to payload if available
-                if (imageUrl != null) {
-                    json.put("image", imageUrl);
-                } else {
-                    Log.d(TAG, "No image URL provided. Image will be null in the database.");
-                }
-
-                // Create request body
-                RequestBody requestBody = RequestBody.create(json.toString(), MediaType.parse("application/json"));
-
-                // Build request
-                Request request = new Request.Builder()
-                        .url(SUPABASE_URL + "/rest/v1/" + SIGNUP_TABLE)
-                        .addHeader("apikey", SUPABASE_KEY)
-                        .addHeader("Authorization", "Bearer " + SUPABASE_KEY)
-                        .post(requestBody)
-                        .build();
-
-                Response response = client.newCall(request).execute();
-
-                if (response.isSuccessful()) {
-                    String responseBody = response.body() != null ? response.body().string() : "No Response Body";
-                    Log.d(TAG, "Signup Success: " + responseBody);
-                    runOnUiThread(() -> {
-                        Toast.makeText(SignUpActivity.this, "Sign up successful!", Toast.LENGTH_SHORT).show();
-                        Intent intent = new Intent(SignUpActivity.this, LoginActivity.class);
-                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                        startActivity(intent);
-                        finish();
-                    });
-                } else {
-                    String errorMessage = response.body() != null ? response.body().string() : "Unknown Error";
-                    Log.e(TAG, "Signup Failed: " + errorMessage);
-
-                    if (errorMessage.contains("duplicate key value violates unique constraint")) {
-                        runOnUiThread(() -> {
-                            Toast.makeText(SignUpActivity.this, "Username taken. Please choose a different one.", Toast.LENGTH_LONG).show();
-                        });
-                    } else {
-                        runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Signup failed: " + errorMessage, Toast.LENGTH_SHORT).show());
-                    }
-                }
-            } catch (Exception e) {
-                Log.e(TAG, "Signup request failed: " + e.getMessage());
-                runOnUiThread(() -> Toast.makeText(SignUpActivity.this, "Signup failed: " + e.getMessage(), Toast.LENGTH_SHORT).show());
-                e.printStackTrace();
-            }
-        }).start();
-    }
 
     private boolean validateInput(String name, String email, String password) {
         if (TextUtils.isEmpty(name)) {
